pldm: Remove unnecessary type casting in logs
This commit removes the unnecessary type casting of values of the
journal logs.

+++ utilities/requester/set_state_effecter.cpp
@@ -53,8 +53,7 @@ int main(int argc, char** argv)
         return -1;
     }
     pldm_msg* response = reinterpret_cast<pldm_msg*>(responseMsg);
-    info("Done. PLDM RC = {RC}", "RC", lg2::hex,
-         static_cast<uint16_t>(response->payload[0]));
+    info("Done. PLDM RC = {RC}", "RC", lg2::hex, response->payload[0]);
     free(responseMsg);
 
     return 0;
